Refactor: MyPostStorage screen
