+++ screens/MyPostStorage/styles.js
@@ -2,8 +2,6 @@ import { StyleSheet } from "react-native";
 
 import {
   WHITE,
-  BLACK,
-  CATEGORY_CARD_TITLE,
   COURSE_COLOR,
   TRANSPARENCY
 } from "../../constants/color";
@@ -43,37 +41,6 @@ const styles = StyleSheet.create({
   },
   loading: {
     backgroundColor: TRANSPARENCY
-  },
-  postContainer: {
-    flexDirection: "row",
-    justifyContent: "space-between"
-  },
-  postTitle: {
-    margin: 10,
-    color: CATEGORY_CARD_TITLE,
-    fontSize: 17,
-    fontWeight: "bold"
-  },
-  postContent: {
-    margin: 10,
-    color: BLACK,
-    fontSize: 15,
-    fontWeight: "600"
-  },
-  likeWrapper: {
-    flexDirection: "row"
-  },
-  likeIcon: {
-    paddingTop: 11
-  },
-  deleteButton: {
-    width: "100%",
-    height: "auto",
-    backgroundColor: TRANSPARENCY,
-  },
-  deleteButtonText: {
-    color: BLACK,
-    fontSize: 17
   },
   emptyContainer: {
     marginTop: "40%",
